requirements.txt
fastapi[all]
python-decouple
PyJWT
bcrypt
pymongo
requests
pydantic
selenium
motor